+++ R/create.cluster.heatmap.R
@@ -1,5 +1,6 @@
 create.cluster.heatmap <- function(
     DF,
+    ccf.limits = NULL,
     clone.colours = NULL,
     height = 6,
     width = 11,
@@ -32,6 +33,14 @@ create.cluster.heatmap <- function(
         xaxis.label <- unique(DF[DF$SNV.id %in% rownames(arr), xaxis.col]);
     } else {
         xaxis.label <- NULL;
+        }
+
+    if (!is.null(ccf.limits)) {
+        if (length(ccf.limits) != 2) {
+            stop('ccf.limits must be a vector of length 2');
+            }
+        arr[arr < ccf.limits[1]] <- ccf.limits[1];
+        arr[arr > ccf.limits[2]] <- ccf.limits[2];
         }
 
     hm <- create.ccf.heatmap(